feat(AED-PGL): some work on matriculas and new android proyect

+++ AED/java/Spring/matriculas/src/main/java/es/iespto/agl/matriculas/service/AlumnoService.java
@@ -1,11 +1,14 @@
 package es.iespto.agl.matriculas.service;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 
 import es.iespto.agl.matriculas.entity.Alumno;
 import es.iespto.agl.matriculas.repository.AlumnoJPARepository;
+import es.iespto.agl.matriculas.repository.MatriculaJPARepository;
+import jakarta.transaction.Transactional;
 
 public class AlumnoService implements IGenericService<Alumno, String>{
 	@Autowired
@@ -18,18 +21,55 @@ public class AlumnoService implements IGenericService<Alumno, String>{
 
 	@Override
 	public Optional<Alumno> findById(String id) {
-		return alumnoRepository.findById(id);
-	}
-
-	@Override
-	public Alumno save(Alumno element) {
-		return alumnoRepository.save(element);
-	}
-
-	@Override
-	public void deleteById(String id) {
-		alumnoRepository.deleteById(id);
+		Optional<Alumno> findById = null;
 		
+		if(id != null) {
+			findById = alumnoRepository.findById(id);
+		}
+	
+		return findById;
+	}
+
+	@Override
+	@Transactional
+	public Alumno save(Alumno element) {
+		Alumno save = null;
+		if(element != null) {
+			if(element.getDni() != null) {
+				save = alumnoRepository.save(element);
+			}
+		}
+		
+		return save;
+	}
+	
+	@Transactional
+	public boolean updateNative(Alumno element) {
+		boolean ok = false;
+		if(element != null) {
+			if(element.getDni() != null) {
+				int updateNtive = alumnoRepository.updateNtive(
+							element.getDni(),
+							element.getApellidos(),
+							element.getFechanacimiento(),
+							element.getFoto(),
+							element.getNombre()
+						);
+			}
+		}
+				
+				ok = true;
+		
+		return ok;
+	}
+
+	@Override
+	@Transactional
+	public void deleteById(String id) {
+		if(id != null) {
+			Optional<Alumno> findById = alumnoRepository.findById(id);
+			alumnoRepository.deleteById(id);
+		}
 	}
 
 }
